youtube_rag_backend/requirements.txt
fastapi
uvicorn
youtube-transcript-api
langchain
langchain_huggingface
langchain_community
faiss-cpu
sentence-transformers
python-dotenv
tiktoken
protobuf
sentencepiece
